requirements.txt
openai==0.27.0
pyttsx3==2.90
SpeechRecognition==3.8.1
google-cloud-aiplatform >= 1.31.0
python-dotenv